+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: gostr <gostr@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/08 12:43:26 by gostr             #+#    #+#             */
-/*   Updated: 2023/11/14 15:12:27 by gostr            ###   ########.fr       */
+/*   Updated: 2023/11/16 13:26:30 by gostr            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t    ft_strlcpy(char *src, char *dest, size_t size)
 {
-    int i;
+    size_t i;
 
     i = 0;
     while (src[i] && i < size)
